editoast: isolate search engine in a dedicated crate
- Moves everything in `editoast_search` expect the endpoint.
- Adapt paths generated by the `Search` derive macro.
- Changes functions returning an editoast custom `Result` to return
    a standard one instead.
- Forwards search engine errors up to the API level, with updated i18n
  keys.
- Merges `views/search/mod.rs` and `views/search/objects.rs` into
  `views/search.rs`.
