+++ editoast/editoast_search/src/search_object.rs
