+++ editoast/editoast_search/src/sql/insert_trigger_template.sql
