+++ editoast/editoast_search/src/sql/update_trigger_template.sql
